ai_agents/requirements.txt
openai==1.54.0
python-dotenv==1.0.0
pydantic==2.9.0
langchain==0.3.0
langchain-openai==0.2.0
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pypdf2==3.0.1
python-docx==1.1.0
geopy==2.4.1
requests==2.32.0
numpy==1.26.0
pandas==2.2.0
aiohttp==3.10.0
